[refactor] 모임 목록 컴포넌트를 공통 컴포넌트로 분리
